# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 2026.5.16-1108mdt
  description: QuantumShield — Post-quantum encrypted chat API (ML-KEM-1024 + ML-DSA-87)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms/{roomId}/messages:
    get:
      operationId: getRoomsRoomIdMessages
      tags: [messages]
      summary: Get encrypted messages in a room
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: before
          in: query
          required: false
          schema:
            type: string
          description: Cursor for pagination (message ID)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of encrypted messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: postRoomsRoomIdMessages
      tags: [messages]
      summary: Send an encrypted message
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /rooms/{roomId}/messages/{messageId}:
    delete:
      operationId: deleteRoomsRoomIdMessagesMessageId
      tags: [messages]
      summary: Delete a message (triggers key destruction)
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Message deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    MessageDecayAttestation:
      type: object
      required:
        - version
        - sampledAt
        - serverFallbackEpochMs
        - synthesizedAt
        - synthesizedEpochMs
        - synthesisMode
        - quorumSize
        - quorumSpreadMs
        - minQuorumSources
        - sources
        - degraded
        - signature
      properties:
        version:
          type: integer
          enum: [1]
        sampledAt:
          type: string
          format: date-time
        serverFallbackEpochMs:
          type: number
        synthesizedAt:
          type: string
          format: date-time
        synthesizedEpochMs:
          type: number
        synthesisMode:
          type: string
          enum: [quorum_median, server_fallback]
          description: Public-source quorum median when enough sources agree, otherwise server fallback.
        quorumSize:
          type: integer
        quorumSpreadMs:
          type: number
        minQuorumSources:
          type: integer
        sources:
          type: array
          items:
            type: object
            required: [id, url, ok]
            properties:
              id:
                type: string
              url:
                type: string
              ok:
                type: boolean
              epochMs:
                type: number
        degraded:
          type: boolean
        signature:
          type: string
          description: Server HMAC over the quorum clock attestation payload.
    Message:
      type: object
      required: [id, roomId, senderId, ciphertext, nonce, algorithm, createdAt]
      properties:
        id:
          type: string
        roomId:
          type: string
        senderId:
          type: string
        senderUsername:
          type: string
          nullable: true
        ciphertext:
          type: string
          description: Base64-encoded AES-256-GCM ciphertext
        nonce:
          type: string
          description: Base64-encoded 96-bit nonce
        algorithm:
          type: string
          description: "Cipher suite: AES-256-GCM+ML-KEM-1024+ML-DSA-87"
        signature:
          type: string
          nullable: true
          description: Base64-encoded ML-DSA-87 signature over ciphertext
        senderDsaPublicKey:
          type: string
          nullable: true
          description: Base64-encoded ML-DSA-87 public key used to verify this message signature.
        recipientEncryptedKeys:
          type: object
          nullable: true
          description: Map of userId to their ML-KEM encapsulated message key
          additionalProperties:
            type: string
        decayAttestation:
          $ref: "#/components/schemas/MessageDecayAttestation"
          nullable: true
          description: Server-signed quorum clock attestation for experimental decay rooms.
        decayedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        availableAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    SendMessageRequest:
      type: object
      required: [ciphertext, nonce, algorithm]
      properties:
        ciphertext:
          type: string
        nonce:
          type: string
        algorithm:
          type: string
        signature:
          type: string
          nullable: true
        recipientEncryptedKeys:
          type: object
          nullable: true
          description: Map of userId to their ML-KEM encapsulated message key
          additionalProperties:
            type: string
        ttlSeconds:
          type: integer
          nullable: true
...