# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuantumShield — Post-quantum encrypted chat API (ML-KEM-1024 + ML-DSA-87)
 * OpenAPI spec version: 2026.5.16-1108mdt
 */
import type { MessageRecipientEncryptedKeys } from './messageRecipientEncryptedKeys';

export interface MessageDecayAttestationSource {
  id: string;
  url: string;
  ok: boolean;
  epochMs?: number;
}

export interface MessageDecayAttestation {
  version: 1;
  sampledAt: string;
  serverFallbackEpochMs: number;
  synthesizedAt: string;
  synthesizedEpochMs: number;
  synthesisMode: 'quorum_median' | 'server_fallback';
  quorumSize: number;
  quorumSpreadMs: number;
  minQuorumSources: number;
  sources: MessageDecayAttestationSource[];
  degraded: boolean;
  signature: string;
}

export interface Message {
  id: string;
  roomId: string;
  senderId: string;
  senderUsername?: string | null;
  /** Base64-encoded AES-256-GCM ciphertext */
  ciphertext: string;
  /** Base64-encoded 96-bit nonce */
  nonce: string;
  /** Cipher suite: AES-256-GCM+ML-KEM-1024+ML-DSA-87 */
  algorithm: string;
  /** Base64-encoded ML-DSA-87 signature over ciphertext */
  signature?: string | null;
  /** Base64-encoded ML-DSA-87 public key used to verify this message signature. */
  senderDsaPublicKey?: string | null;
  /** Map of userId to their ML-KEM encapsulated message key */
  recipientEncryptedKeys?: MessageRecipientEncryptedKeys;
  /** Server-signed quorum clock attestation for experimental decay rooms. */
  decayAttestation?: MessageDecayAttestation | null;
  decayedAt?: Date | null;
  expiresAt?: Date | null;
  availableAt?: Date | null;
  createdAt: Date;
}